extract RenderError
