+++ packages/insomnia/src/network/network.ts
@@ -34,6 +34,7 @@ import type { WebSocketRequest } from '../models/websocket-request';
 import { isWorkspace, type Workspace } from '../models/workspace';
 import * as pluginContexts from '../plugins/context/index';
 import * as plugins from '../plugins/index';
+import { RenderError } from '../templating/render-error';
 import { maskOrDecryptContextIfNecessary } from '../templating/utils';
 import { defaultSendActionRuntime, type SendActionRuntime } from '../ui/routes/request';
 import { invariant } from '../utils/invariant';
@@ -618,17 +619,23 @@ export const tryToInterpolateRequest = async ({
   ignoreUndefinedEnvVariable?: boolean;
 }
 ) => {
-  return await getRenderedRequestAndContext({
-    request: request,
-    environment,
-    baseEnvironment,
-    userUploadEnvironment,
-    transientVariables,
-    purpose,
-    extraInfo,
-    ignoreUndefinedEnvVariable,
-  });
-
+  try {
+    return await getRenderedRequestAndContext({
+      request: request,
+      environment,
+      baseEnvironment,
+      userUploadEnvironment,
+      transientVariables,
+      purpose,
+      extraInfo,
+      ignoreUndefinedEnvVariable,
+    });
+  } catch (err) {
+    if (err instanceof RenderError) {
+      throw err;
+    }
+    throw new Error(`Failed to render request: ${request._id}`);
+  }
 };
 
 export const tryToTransformRequestWithPlugins = async (renderResult: RequestAndContext) => {
