+++ packages/insomnia/src/utils/try-interpolate.ts
@@ -2,6 +2,7 @@ import { getRenderContext, render } from '../common/render';
 import type { GrpcRequest } from '../models/grpc-request';
 import type { Request } from '../models/request';
 import type { WebSocketRequest } from '../models/websocket-request';
+import { RenderError } from '../templating/render-error';
 import { showModal } from '../ui/components/modals';
 import { RequestRenderErrorModal } from '../ui/components/modals/request-render-error-modal';
 
@@ -11,7 +12,7 @@ export const tryToInterpolateRequestOrShowRenderErrorModal = async ({ request, e
     const renderContext = await getRenderContext({ request, environment: environmentId, purpose: 'send' });
     return await render(payload, renderContext);
   } catch (error) {
-    if (error.type === 'render') {
+    if (error instanceof RenderError) {
       showModal(RequestRenderErrorModal, { request, error });
       return;
     }
